Add new Section storybook doc
