--- packages/storybook/examples/core/Section/DemoContent.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const styles = {
-    width: '100%',
-    lineHeight: '200px',
-    textAlign: 'center',
-    background: '#eee',
-};
-
-function DemoContent() {
-    return (
-        <div style={styles}>
-            Section Content
-        </div>
-    );
-}
-
-export default DemoContent;
